Refactor base packages
