+++ atrium/speculatio/fenestra/main.go
@@ -7,15 +7,8 @@ package main
 import (
 	"embed"
 	"flag"
-	"log"
-	"os"
 
-	"github.com/trimble-oss/tierceron/atrium/speculatio/fenestra/data"
-	"github.com/trimble-oss/tierceron/atrium/speculatio/fenestra/trcRenderers"
-
-	"fyne.io/fyne/v2"
-	"github.com/trimble-oss/tierceron-nute/custos/custosworld"
-	"github.com/trimble-oss/tierceron-nute/mashupsdk"
+	"github.com/trimble-oss/tierceron/atrium/speculatio/fenestra/fenestrabase"
 )
 
 var worldCompleteChan chan bool
@@ -25,10 +18,6 @@ var mashupCert embed.FS
 
 //go:embed tls/mashup.key
 var mashupKey embed.FS
-
-func OutsideClone(custosWorldApp *custosworld.CustosWorldApp, childId int64, concreteElement *mashupsdk.MashupDetailedElement) {
-	custosWorldApp.FyneWidgetElements["Outside"].MashupDetailedElement.Copy(concreteElement)
-}
 
 //go:embed logo.png
 var logoIcon embed.FS
@@ -41,44 +30,12 @@ func main() {
 	envPtr := flag.String("env", "QA", "Environment to configure")
 	flag.Parse()
 
-	helloLog, err := os.OpenFile("fenestra.log", os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0600)
-	if err != nil {
-		log.Fatalf(err.Error(), err)
-	}
-	logger := log.New(helloLog, "[fenestra]", log.LstdFlags)
-	log.SetOutput(helloLog)
-
-	mashupsdk.InitCertKeyPair(mashupCert, mashupKey)
-	var DetailedElements []*mashupsdk.MashupDetailedElement
-
-	if *headless {
-		DetailedElements = data.GetHeadlessData(insecure, logger)
-	} else {
-		DetailedElements = data.GetData(insecure, logger, envPtr)
-	}
-	if len(DetailedElements) > 0 {
-		logger.Printf("Successfully loaded %d elements.\n", len(DetailedElements))
-	} else {
-		logger.Printf("Failure to load any enterprises.\n")
-	}
-
-	tenantDataRenderer := &trcRenderers.TenantDataRenderer{}
-	custosWorld := custosworld.NewCustosWorldApp(*serverheadless, false, DetailedElements, tenantDataRenderer)
-	tenantDataRenderer.CustosWorldApp = custosWorld
-	custosWorld.CustomTabItemRenderer["TenantDataRenderer"] = tenantDataRenderer
-	custosWorld.CustomTabItemRenderer["SearchRenderer"] = &trcRenderers.SearchRenderer{CustosWorldApp: custosWorld}
-
-	custosWorld.Title = "Tierceron Topology Discovery Interface"
-	custosWorld.MainWindowSize = fyne.NewSize(800, 100)
-	logoIconBytes, _ := logoIcon.ReadFile("logo.png")
-	custosWorld.Icon = fyne.NewStaticResource("Logo", logoIconBytes)
-
-	if !custosWorld.Headless {
-		custosWorld.InitServer(*callerCreds, *insecure, 500*1024*1024)
-	}
-
-	// Initialize the main window.
-	custosWorld.InitMainWindow()
-
-	<-worldCompleteChan
+	fenestrabase.CommonMain(logoIcon,
+		mashupCert,
+		mashupKey,
+		callerCreds,
+		insecure,
+		headless,
+		serverheadless,
+		envPtr)
 }
